内部协同修改

+++ src/views/synergy/index.coffee
@@ -3,48 +3,74 @@ import Utils from '@/assets/scripts/utils'
 
 export default
 	data: ->
-		dsialogFormVisible:false
-		indexThis:-1
-		alertTitle:''
-		initModel:''
-		alertType:''
-		resdata:[]
-		adminsList:''
-		peopleCount:'--'
-		groupsList:[]
-		loadedData:false
-		disbaledFont:'禁用'
-		initIndex:0
-		servenumber:0
-		kefnumber:0
-		disablednumber:0
+		# 窗口是否显示
+		dsialogFormVisible: false
+		# 样式切换
+		indexThis: -1
+		# 窗口标题
+		alertTitle: ''
+		# 窗口input值
+		initModel: ''
+		# 窗口类型？修改or新增
+		alertType: ''
+		# 初始化数据赋值
+		resdata: []
+		groupsList: []
+		adminsList: ''
+		# 被禁用人数值
+		peopleCount: '--'
+		# 加载完成
+		loadedData: false
+		disbaledFont: '禁用'
+		inputmsg: '请输入分组名称'
+		titleinfo: '确定禁用该客服？'
+		initIndex: 0
+		servenumber: 0
+		kefnumber: 0
+		disablednumber: 0
+		errorTipbox: false
+		# 防止多次点击
+		isxhr: false
+		isclick: false
+		# 是否含有input框的窗口
+		hasInput: true
+		# 点击对应的ID值
+		operatingId: ''
+		# 启动或禁用 1 => 禁用 2 => 启动
+		status_init: ''
+		# 分组输入框正则
+		gobaleReg: /^[\u4e00-\u9fa5_a-zA-Z0-9`~!@#$%^&*()_\-+=?:"{}|,.\/;'\\[\]·~！@#￥%……&*（）——\-+={}|《》？：“”【】、；‘’，。、]{1,20}$/
 	created:->
 		this.loadAll()
+	watch:
+		$route:(to)->
+			this.loadAll() if to.name is 'synergy'
 	methods:
+		# 点击隐藏错误提示
+		hiddentip:()->
+			this.errorTipbox=false
+			this.$refs.inputBox.focus()
+			this.inputmsg = '请输入分组名称'
 		# 删除客服
 		deletedServe:(id)->
-			Utils.ajax ALPHA.API_PATH.synergy.deletedServe,
-				method: 'delete'
-				params: id:id
-			.then (res) =>
-				this.loadAll()
+			this.hasInput = false
+			this.dsialogFormVisible = true
+			this.titleinfo = "确定删除该客服？"
+			this.operatingId = id
 		# 禁用or启用客服
 		disabeldServe:(event,index,id)->
-			status_init = ''
+			this.hasInput = false
 			htmldom =  event.target.innerHTML
+			this.operatingId = id
 			if htmldom is '禁用'
-				event.target.innerHTML = '启用'
-				status_init = 1
+				this.titleinfo = "确定启用该客服？"
+				this.status_init = 1
+				this.dsialogFormVisible = true
 			else if htmldom is '启用'
-				status_init = 2
-				event.target.innerHTML = '禁用'
-			Utils.ajax ALPHA.API_PATH.synergy.disabledServe,
-				method: 'put'
-				data:
-					id:id
-					status:status_init
-			.then (res) =>
-				this.loadAll()
+				this.status_init = 2
+				this.titleinfo = "确定禁用该客服？"
+				this.dsialogFormVisible = true
+			
 		# 初始化数据
 		loadAll:->
 			Utils.ajax ALPHA.API_PATH.synergy.all
@@ -55,8 +81,9 @@ export default
 				this.peopleCount = this.adminsList.length
 				this.groundList(this.adminsList,this.resdata.groups)
 		# 跳转到添加客服
-		aHerf:->
-			this.$router.push({path:'/synergy/addService'})
+		aHerf:(id)->
+			this.$router.push({path:'/synergy/addService',query:{account:id}}) if id 
+			this.$router.push({path:'/synergy/addService'}) unless id 
 		groundList:(adminsList,groupsList)->
 			for grounpItems,grounpI in groupsList
 				groupsList[grounpI].newVal = []
@@ -66,7 +93,7 @@ export default
 					else
 						if adminsItem.groupId is grounpItems.id
 							groupsList[grounpI].newVal.push(adminsItem)
-			this.loadRightData(0)
+			this.loadRightData(this.initIndex)
 			this.loadedData = true
 		# 添加分组
 		addGrouping: ->
@@ -76,16 +103,53 @@ export default
 			this.dsialogFormVisible = true
 		# 关闭窗口
 		canleAlert:->
+			this.hasInput = true
+			this.errorTipbox = false
+			this.inputmsg = '请输入分组名称'
 			this.dsialogFormVisible = false
 		# 删除分组
 		deleteGround:(id)->
-			Utils.ajax ALPHA.API_PATH.synergy.deleted,
+			this.hasInput = false
+			this.dsialogFormVisible = true
+			this.titleinfo = "确定删除该分组？"
+			this.operatingId = id
+		# 确定操作
+		determine:(msg)->
+			return no if this.isclick is on
+			this.isclick = true
+			if msg.indexOf('删除该分组') >= 1
+				Utils.ajax ALPHA.API_PATH.synergy.deleted,
 					method: 'delete'
-					params: id:id
+					params: id:this.operatingId
 				.then (res) =>
-					this.loadAll()
+					this.isclick = false
+			else if msg.indexOf('删除该客服') >= 1
+				Utils.ajax ALPHA.API_PATH.synergy.deletedServe,
+					method: 'delete'
+					params: id:this.operatingId
+				.then (res) =>
+					this.isclick = false
+			else if msg.indexOf("启用") >= 1 or msg.indexOf("禁用") >= 1
+				Utils.ajax ALPHA.API_PATH.synergy.disabledServe,
+					method: 'put'
+					data:
+						id:this.operatingId
+						status:this.status_init
+				.then (res) =>
+					this.isclick = false
+			this.canleAlert()
+			this.loadAll()
 		# 弹窗数据保存
 		saveGround:->
+			unless this.initModel
+				this.errorTipbox=true
+				this.inputmsg = ''
+				return no
+			else if this.gobaleReg.test(this.initModel)
+				this.errorTipbox=true
+				this.inputmsg = ''
+				return no
+			this.isxhr = true
 			if this.alertType is 1 
 				Utils.ajax ALPHA.API_PATH.synergy.group,
 					method: 'POST'
@@ -93,6 +157,7 @@ export default
 				.then (res) =>
 					this.canleAlert()
 					this.loadAll()
+					this.isxhr = false
 			else if this.alertType is 2
 				Utils.ajax ALPHA.API_PATH.synergy.edit,
 					method: 'PUT'
@@ -102,6 +167,8 @@ export default
 				.then (res) =>
 					this.canleAlert()
 					this.loadAll()
+					this.isxhr = false
+
 		# css样式切换
 		bMouseEnterFun:(index)->
 			this.indexThis = index
@@ -115,7 +182,6 @@ export default
 			this.disablednumber = 0
 			this.initIndex = index
 			for item in  this.groupsList[index].newVal 
-				console.log item.roleId
 				# item.roleId == 1 ？ 管理员 ：客服
 				if item.roleId is 1
 					this.servenumber += 1
@@ -124,7 +190,6 @@ export default
 				# 被禁用的人数
 				if item.status is 1
 					this.disablednumber += 1
-			console.log("点击加载数据")
 		# 修改分组名称
 		editGround:(itemsName,id)->
 			this.initModel = itemsName
